+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Changelog
+
+## 2.0.0
+
+- Now uses its own api to fetch emoji data
+- generate tags and categories for emojis
+
 ## 1.0.0
 
 - Initial version.